smaller tests for CI/CD
